created entities, enums, changed application.properties default scheme database

+++ src/main/java/kg/bank/payments/model/entity/ServiceDetail.java
@@ -32,7 +32,7 @@ public class ServiceDetail {
     Account account;
 
     @Column(name = "percent_price")
-    Double percentPrice;
+    BigDecimal percentPrice;
 
     @Column(name = "fix_price")
     BigDecimal fixPrice;
